Add ability to fade in/out of the first/last track
Live recordings often have noise till the end or need to have a jingle
cut. Use this to create a smooth fade-in into the first track or a
fade-out of the last one. Always applies to the first/last track of a
snipfile, not affected by track selection via --tracks.